app/requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
SQLAlchemy==2.0.32
python-dotenv==1.0.1
nltk==3.9.1
openai==1.51.2
beautifulsoup4==4.12.3
